move entity stuff in one package
