+++ blockly/src/level/produs/Level015.java
@@ -11,7 +11,7 @@ import core.level.utils.DesignLabel;
 import core.level.utils.LevelElement;
 import core.utils.Direction;
 import core.utils.MissingHeroException;
-import entities.BlocklyMonster;
+import entities.monster.BlocklyMonster;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
